Fix np type bug and add regression test

+++ calibration_environment/drivers/water_bath.py
@@ -346,7 +346,7 @@ def _construct_command_packet(command_name: str, data: float = None):
 
         data_byte_count = 2  # Data is always sent as two bytes
 
-        data_bytes = round(data / REPORTING_PRECISION).to_bytes(
+        data_bytes = int(round(data / REPORTING_PRECISION)).to_bytes(
             data_byte_count, byteorder="big"
         )
